requirements.txt
# Dependências principais
python-telegram-bot==20.7
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.2
pydantic-settings==2.1.0
pytest-asyncio==0.21.1

# OCR e processamento de imagens
pytesseract==0.3.10
Pillow==10.1.0

# Visualização de dados
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.4

# Monitoramento
psutil==5.9.7

# Agendamento
schedule==1.2.1

# Utilitários
python-dateutil==2.8.2
python-dotenv==1.0.0

fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
aiohttp==3.9.1
tenacity==8.2.2 